+++ odoo/local-src/specific_report/models/report_async.py
@@ -2,7 +2,7 @@
 # © 2018 Camptocamp SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-from odoo import api, models
+from odoo import api, models, _
 from odoo.addons.queue_job.job import job
 
 
@@ -15,8 +15,12 @@ class ReportAsync(models.AbstractModel):
 
     @job(default_channel='root.background.report')
     @api.multi
-    def print_and_attach_report(self, report):
-        """Print and attach a report"""
+    def print_and_attach_report(self, report, send_to_fax=None):
+        """Print and attach a report.
+
+        param send_to_fax: If not empty the attachment will be sent
+        to the number specified.
+        """
         self.ensure_one()
         filename = self.get_report_name()
         data = self.env['report'].get_pdf([self.id], report)
@@ -25,7 +29,7 @@ class ReportAsync(models.AbstractModel):
         if len(existing):
             existing[0].datas = data.encode('base_64')
         else:
-            self.env['ir.attachment'].create({
+            new_report = self.env['ir.attachment'].create({
                 'type': 'binary',
                 'res_model': self._name,
                 'res_id': self.id,
@@ -34,3 +38,11 @@ class ReportAsync(models.AbstractModel):
                 'mimetype': 'application/pdf',
                 'datas': data.encode('base_64'),
                 })
+        if send_to_fax:
+            report_id = existing[0].id if len(existing) else new_report.id
+            fax = self.env.ref('external_fax.ovh')
+            fax.with_delay(
+                description=_(
+                    'Sending fax for {} with id {}'
+                ).format(self._name, self.id),
+            ).send(send_to_fax, report_id)