+++ odoo/local-src/specific_report/models/sale_order.py
@@ -50,8 +50,8 @@ class SaleOrder(models.Model):
     def get_report_name(self):
         """Generate a specific name for the report save in ir.attachment"""
         self.ensure_one()
-        if self.state == 'draft':
-            # Not saving in ir.attachment when in draft state
+        if self.state not in ['sale', 'done']:
+            # Not saving the report in ir.attachment, when not confirmed
             return None
         if not self.partner_id.ref:
             raise UserError(_(