+++ HISTORY.rst
@@ -25,6 +25,7 @@ Unreleased
   * Display lots on product.product and product.template
   * Fix a bug on the query to compute lot to archive. No lots was archived
   * On lots change ref to product quantity
+* ALCYN-486: When creating a new sale order coming from the web service, make sure it is set with the appropriate default values (calling the existing onchange methods)
 
 **Build**